refactor: split BlogPage (1473 lines) and LoginPage (1098 lines) into modules
BlogPage:
- Extract articles.ts — static article data (~920 lines)
- Extract ArticleModal.tsx — article detail modal with markdown rendering
- Extract CommentSection.tsx — comments with CRUD operations
- index.tsx reduced to ~200 lines (list + filter + layout)

LoginPage:
- Extract ForgotPasswordModal.tsx — self-contained password reset modal
- Extract GDPRModal.tsx — GDPR compliance info modal with static data
- Fix error: any patterns in auth handlers (catch unknown)
- index.tsx reduced to ~575 lines (auth logic + main form)
