+++ lexwebapp/src/pages/BlogPage/articles.ts
@@ -1,31 +1,8 @@
-import { useState, useEffect, useCallback } from 'react';
-import { motion, AnimatePresence } from 'framer-motion';
-import { useNavigate, useSearchParams } from 'react-router-dom';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import {
-  ArrowLeft,
-  BookOpen,
-  Linkedin,
-  Link2,
-  Check,
-  ChevronRight,
-  Cpu,
-  Scale,
-  X,
-  MessageSquare,
-  Send,
-  Trash2,
-  LogIn,
-  Facebook,
-  Twitter,
-} from 'lucide-react';
-import { useAuth } from '../contexts/AuthContext';
-import { blogService, type BlogComment } from '../services/blog-service';
+/**
+ * Blog article data — static content for the blog page
+ */
 
-// ── Article data ──────────────────────────────────────────────
-
-interface Article {
+export interface Article {
   id: string;
   title: string;
   punchline: string;
@@ -35,7 +12,7 @@ interface Article {
   content: string;
 }
 
-const articles: Article[] = [
+export const articles: Article[] = [
   {
     id: 'attorney-marketplace',
     title: 'Маркетплейс юридичних консультацій: від реєстру ЄРАУ до оплати через Monobank',
@@ -955,519 +932,3 @@ AI не вирішує, яку стратегію обрати. Він дає ю
 Конфіденційність — не фіча. Це передумова існування юридичної AI-платформи.`,
   },
 ];
-
-// ── Component ─────────────────────────────────────────────────
-
-export function BlogPage() {
-  const navigate = useNavigate();
-  const [searchParams, setSearchParams] = useSearchParams();
-  const { user, isAuthenticated } = useAuth();
-  const [selectedArticle, setSelectedArticle] = useState<Article | null>(null);
-  const [copied, setCopied] = useState(false);
-  const [filter, setFilter] = useState<'all' | 'tech' | 'legal'>('all');
-
-  // Auto-open article from ?article= query param (used after login redirect)
-  useEffect(() => {
-    const articleId = searchParams.get('article');
-    if (articleId && !selectedArticle) {
-      const found = articles.find((a) => a.id === articleId);
-      if (found) {
-        setSelectedArticle(found);
-        // Clean the query param from URL
-        searchParams.delete('article');
-        setSearchParams(searchParams, { replace: true });
-      }
-    }
-  }, [searchParams]);
-
-  // Comments state
-  const [comments, setComments] = useState<BlogComment[]>([]);
-  const [commentText, setCommentText] = useState('');
-  const [commentsLoading, setCommentsLoading] = useState(false);
-  const [commentSubmitting, setCommentSubmitting] = useState(false);
-
-  const loadComments = useCallback(async (articleId: string) => {
-    setCommentsLoading(true);
-    try {
-      const data = await blogService.getComments(articleId);
-      setComments(data);
-    } catch {
-      setComments([]);
-    } finally {
-      setCommentsLoading(false);
-    }
-  }, []);
-
-  // Load comments when article is selected
-  useEffect(() => {
-    if (selectedArticle) {
-      loadComments(selectedArticle.id);
-      setCommentText('');
-    } else {
-      setComments([]);
-    }
-  }, [selectedArticle, loadComments]);
-
-  const handleSubmitComment = async () => {
-    if (!selectedArticle || !commentText.trim() || commentSubmitting) return;
-    setCommentSubmitting(true);
-    try {
-      const newComment = await blogService.postComment(selectedArticle.id, commentText.trim());
-      setComments(prev => [newComment, ...prev]);
-      setCommentText('');
-    } catch {
-      // handled silently
-    } finally {
-      setCommentSubmitting(false);
-    }
-  };
-
-  const handleDeleteComment = async (commentId: number) => {
-    try {
-      await blogService.deleteComment(commentId);
-      setComments(prev => prev.filter(c => c.id !== commentId));
-    } catch {
-      // handled silently
-    }
-  };
-
-  const filtered = filter === 'all' ? articles : articles.filter(a => a.category === filter);
-
-  const getArticleUrl = (article: Article) =>
-    `${window.location.origin}/blog?article=${article.id}`;
-
-  const shareOnLinkedIn = (article: Article) => {
-    const url = encodeURIComponent(getArticleUrl(article));
-    const text = encodeURIComponent(`${article.punchline}\n\n#LegalTech #AI #LEXai`);
-    window.open(`https://www.linkedin.com/sharing/share-offsite/?url=${url}&text=${text}`, '_blank');
-  };
-
-  const shareOnX = (article: Article) => {
-    const url = encodeURIComponent(getArticleUrl(article));
-    const text = encodeURIComponent(`${article.title}\n\n#LegalTech #AI`);
-    window.open(`https://x.com/intent/tweet?url=${url}&text=${text}`, '_blank');
-  };
-
-  const shareOnFacebook = (article: Article) => {
-    const url = encodeURIComponent(getArticleUrl(article));
-    window.open(`https://www.facebook.com/sharer/sharer.php?u=${url}`, '_blank');
-  };
-
-  const copyLink = (article: Article) => {
-    navigator.clipboard.writeText(getArticleUrl(article));
-    setCopied(true);
-    setTimeout(() => setCopied(false), 2000);
-  };
-
-  return (
-    <div className="min-h-screen bg-claude-bg">
-      {/* Header */}
-      <header className="sticky top-0 z-20 bg-white/90 backdrop-blur-md border-b border-claude-border">
-        <div className="max-w-5xl mx-auto px-4 sm:px-6 py-4 flex items-center justify-between">
-          <div className="flex items-center gap-4">
-            <button
-              onClick={() => navigate('/login')}
-              className="flex items-center gap-2 text-claude-subtext hover:text-claude-text transition-colors font-sans text-sm"
-            >
-              <ArrowLeft size={16} />
-              <span className="hidden sm:inline">LEX AI</span>
-            </button>
-            <div className="h-5 w-px bg-claude-border" />
-            <div className="flex items-center gap-2">
-              <BookOpen size={18} className="text-claude-accent" />
-              <span className="font-sans font-medium text-claude-text text-sm">Blog</span>
-            </div>
-          </div>
-          <a
-            href="/login"
-            className="px-4 py-2 bg-claude-accent text-white rounded-xl text-sm font-medium font-sans hover:bg-[#C66345] transition-colors"
-          >
-            Спробувати LEX AI
-          </a>
-        </div>
-      </header>
-
-      {/* Hero */}
-      <section className="bg-gradient-to-b from-white to-claude-bg border-b border-claude-border">
-        <div className="max-w-5xl mx-auto px-4 sm:px-6 py-12 sm:py-16 text-center">
-          <motion.h1
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            className="text-3xl sm:text-4xl font-serif text-claude-text mb-4"
-          >
-            LEX AI Blog
-          </motion.h1>
-          <motion.p
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ delay: 0.1 }}
-            className="text-claude-subtext font-sans max-w-2xl mx-auto"
-          >
-            Як ми будуємо AI-платформу для юридичної аналітики.
-            Технічні інсайти для розробників та практичні кейси для юристів.
-          </motion.p>
-
-          {/* Filter tabs */}
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            transition={{ delay: 0.2 }}
-            className="flex justify-center gap-2 mt-8"
-          >
-            {([
-              { key: 'all', label: 'Всі', count: articles.length },
-              { key: 'tech', label: 'Tech', count: articles.filter(a => a.category === 'tech').length },
-              { key: 'legal', label: 'Для юристів', count: articles.filter(a => a.category === 'legal').length },
-            ] as const).map(tab => (
-              <button
-                key={tab.key}
-                onClick={() => setFilter(tab.key)}
-                className={`px-4 py-2 rounded-xl text-sm font-sans font-medium transition-all ${
-                  filter === tab.key
-                    ? 'bg-claude-accent text-white shadow-sm'
-                    : 'bg-white text-claude-subtext hover:text-claude-text border border-claude-border'
-                }`}
-              >
-                {tab.label}
-                <span className={`ml-1.5 text-xs ${filter === tab.key ? 'text-white/70' : 'text-claude-subtext/50'}`}>
-                  {tab.count}
-                </span>
-              </button>
-            ))}
-          </motion.div>
-        </div>
-      </section>
-
-      {/* Article Cards */}
-      <section className="max-w-5xl mx-auto px-4 sm:px-6 py-8 sm:py-12">
-        <div className="grid gap-5">
-          {filtered.map((article, i) => (
-            <motion.article
-              key={article.id}
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ delay: i * 0.05 }}
-              className="bg-white rounded-2xl border border-claude-border overflow-hidden hover:shadow-lg hover:border-claude-accent/30 transition-all cursor-pointer group"
-              onClick={() => setSelectedArticle(article)}
-            >
-              <img
-                src={`/blog-banners/${article.id}.png`}
-                alt={article.title}
-                className="w-full h-40 sm:h-48 object-cover"
-              />
-              <div className="p-6 sm:p-8">
-              <div className="flex items-start gap-4">
-                <div className={`flex-shrink-0 w-10 h-10 rounded-xl flex items-center justify-center ${
-                  article.category === 'tech'
-                    ? 'bg-blue-50 text-blue-600'
-                    : 'bg-claude-accent/10 text-claude-accent'
-                }`}>
-                  {article.category === 'tech' ? <Cpu size={20} /> : <Scale size={20} />}
-                </div>
-                <div className="min-w-0 flex-1">
-                  <div className="flex items-center gap-2 mb-2">
-                    <span className={`text-[10px] font-bold tracking-wider uppercase font-sans px-2 py-0.5 rounded-full ${
-                      article.category === 'tech'
-                        ? 'bg-blue-50 text-blue-600'
-                        : 'bg-claude-accent/10 text-claude-accent'
-                    }`}>
-                      {article.category === 'tech' ? 'TECH' : 'LEGAL'}
-                    </span>
-                    <span className="text-xs text-claude-subtext/60 font-sans">{article.readTime}</span>
-                  </div>
-                  <h2 className="text-lg sm:text-xl font-serif text-claude-text mb-3 group-hover:text-claude-accent transition-colors">
-                    {article.title}
-                  </h2>
-                  <p className="text-sm text-claude-subtext font-sans leading-relaxed line-clamp-3">
-                    {article.punchline}
-                  </p>
-                  <div className="flex items-center gap-2 mt-4 flex-wrap">
-                    {article.tags.map(tag => (
-                      <span
-                        key={tag}
-                        className="text-[11px] px-2 py-0.5 rounded-full bg-claude-bg text-claude-subtext font-sans"
-                      >
-                        #{tag}
-                      </span>
-                    ))}
-                    <div className="flex-1" />
-                    <span className="text-sm text-claude-accent font-sans font-medium flex items-center gap-1 opacity-0 group-hover:opacity-100 transition-opacity">
-                      Читати <ChevronRight size={14} />
-                    </span>
-                  </div>
-                </div>
-              </div>
-              </div>
-            </motion.article>
-          ))}
-        </div>
-      </section>
-
-      {/* Footer */}
-      <footer className="border-t border-claude-border bg-white">
-        <div className="max-w-5xl mx-auto px-4 sm:px-6 py-8 flex flex-col sm:flex-row items-center justify-between gap-4">
-          <div className="flex items-center gap-3">
-            <img src="/Image.jpg" alt="Lex" className="h-8 w-auto" />
-            <span className="text-sm text-claude-subtext font-sans">LEX AI — AI-powered legal research platform</span>
-          </div>
-          <div className="flex items-center gap-4 text-sm text-claude-subtext font-sans">
-            <a href="/login" className="hover:text-claude-accent transition-colors">Увійти</a>
-            <span className="text-claude-border">|</span>
-            <a href="/ua/data-sources" className="hover:text-claude-accent transition-colors">Джерела даних</a>
-          </div>
-        </div>
-      </footer>
-
-      {/* Article Modal */}
-      <AnimatePresence>
-        {selectedArticle && (
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            className="fixed inset-0 z-50 bg-black/50 backdrop-blur-sm overflow-y-auto"
-            onClick={(e) => {
-              if (e.target === e.currentTarget) setSelectedArticle(null);
-            }}
-          >
-            <motion.div
-              initial={{ opacity: 0, y: 40 }}
-              animate={{ opacity: 1, y: 0 }}
-              exit={{ opacity: 0, y: 40 }}
-              transition={{ type: 'spring', damping: 30, stiffness: 300 }}
-              className="max-w-3xl mx-auto my-8 sm:my-12 bg-white rounded-2xl shadow-2xl overflow-hidden"
-            >
-              {/* Modal header */}
-              <div className="sticky top-0 z-10 bg-white/95 backdrop-blur-sm border-b border-claude-border px-6 sm:px-8 py-4 flex items-center justify-between">
-                <div className="flex items-center gap-3">
-                  <span className={`text-[10px] font-bold tracking-wider uppercase font-sans px-2 py-0.5 rounded-full ${
-                    selectedArticle.category === 'tech'
-                      ? 'bg-blue-50 text-blue-600'
-                      : 'bg-claude-accent/10 text-claude-accent'
-                  }`}>
-                    {selectedArticle.category === 'tech' ? 'TECH' : 'LEGAL'}
-                  </span>
-                  <span className="text-xs text-claude-subtext font-sans">{selectedArticle.readTime}</span>
-                </div>
-                <div className="flex items-center gap-2">
-                  <button
-                    onClick={(e) => { e.stopPropagation(); shareOnLinkedIn(selectedArticle); }}
-                    className="w-8 h-8 flex items-center justify-center rounded-lg text-[#0A66C2] bg-[#0A66C2]/10 hover:bg-[#0A66C2]/20 transition-colors"
-                    title="Поділитися в LinkedIn"
-                  >
-                    <Linkedin size={15} />
-                  </button>
-                  <button
-                    onClick={(e) => { e.stopPropagation(); shareOnX(selectedArticle); }}
-                    className="w-8 h-8 flex items-center justify-center rounded-lg text-claude-text bg-claude-bg hover:bg-claude-border transition-colors"
-                    title="Поділитися в X"
-                  >
-                    <Twitter size={15} />
-                  </button>
-                  <button
-                    onClick={(e) => { e.stopPropagation(); shareOnFacebook(selectedArticle); }}
-                    className="w-8 h-8 flex items-center justify-center rounded-lg text-[#1877F2] bg-[#1877F2]/10 hover:bg-[#1877F2]/20 transition-colors"
-                    title="Поділитися у Facebook"
-                  >
-                    <Facebook size={15} />
-                  </button>
-                  <button
-                    onClick={(e) => { e.stopPropagation(); copyLink(selectedArticle); }}
-                    className="w-8 h-8 flex items-center justify-center rounded-lg text-claude-subtext bg-claude-bg hover:bg-claude-border transition-colors"
-                    title="Копіювати посилання"
-                  >
-                    {copied ? <Check size={15} className="text-green-600" /> : <Link2 size={15} />}
-                  </button>
-                  <button
-                    onClick={() => setSelectedArticle(null)}
-                    className="w-8 h-8 flex items-center justify-center rounded-lg hover:bg-claude-bg text-claude-subtext hover:text-claude-text transition-colors"
-                  >
-                    <X size={18} />
-                  </button>
-                </div>
-              </div>
-
-              {/* Banner */}
-              <img
-                src={`/blog-banners/${selectedArticle.id}.png`}
-                alt={selectedArticle.title}
-                className="w-full h-48 sm:h-64 object-cover"
-              />
-
-              {/* Modal content */}
-              <div className="px-6 sm:px-8 py-6 sm:py-8">
-                <div className="prose prose-sm sm:prose max-w-none
-                  prose-headings:font-serif prose-headings:text-claude-text
-                  prose-h1:text-2xl sm:prose-h1:text-3xl prose-h1:mb-6
-                  prose-h2:text-xl prose-h2:mt-8 prose-h2:mb-4
-                  prose-h3:text-lg prose-h3:mt-6 prose-h3:mb-3
-                  prose-p:text-claude-text prose-p:font-sans prose-p:leading-relaxed
-                  prose-li:text-claude-text prose-li:font-sans
-                  prose-strong:text-claude-text
-                  prose-em:text-claude-subtext
-                  prose-code:text-claude-accent prose-code:bg-claude-accent/5 prose-code:px-1.5 prose-code:py-0.5 prose-code:rounded prose-code:text-sm prose-code:font-normal prose-code:before:content-none prose-code:after:content-none
-                  prose-pre:bg-gray-900 prose-pre:rounded-xl
-                  prose-blockquote:border-claude-accent prose-blockquote:bg-claude-accent/5 prose-blockquote:rounded-r-xl prose-blockquote:py-1
-                  prose-a:text-claude-accent prose-a:no-underline hover:prose-a:underline
-                  prose-table:font-sans
-                  prose-th:bg-claude-bg prose-th:text-claude-text prose-th:font-medium prose-th:text-sm prose-th:px-4 prose-th:py-2.5
-                  prose-td:text-sm prose-td:px-4 prose-td:py-2.5 prose-td:border-claude-border
-                  prose-hr:border-claude-border
-                ">
-                  <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                    {selectedArticle.content}
-                  </ReactMarkdown>
-                </div>
-
-                {/* Tags */}
-                <div className="flex flex-wrap gap-2 mt-8 pt-6 border-t border-claude-border">
-                  {selectedArticle.tags.map(tag => (
-                    <span
-                      key={tag}
-                      className="text-xs px-2.5 py-1 rounded-full bg-claude-bg text-claude-subtext font-sans"
-                    >
-                      #{tag}
-                    </span>
-                  ))}
-                </div>
-
-                {/* Share footer */}
-                <div className="mt-6 p-4 bg-claude-bg rounded-xl flex flex-col sm:flex-row items-center justify-between gap-4">
-                  <p className="text-sm text-claude-subtext font-sans">
-                    Поділитися цією статтею:
-                  </p>
-                  <div className="flex items-center gap-2">
-                    <button
-                      onClick={() => shareOnLinkedIn(selectedArticle)}
-                      className="flex items-center gap-2 px-4 py-2.5 bg-[#0A66C2] text-white rounded-xl text-sm font-sans font-medium hover:bg-[#004182] transition-colors"
-                    >
-                      <Linkedin size={16} />
-                      LinkedIn
-                    </button>
-                    <button
-                      onClick={() => shareOnX(selectedArticle)}
-                      className="flex items-center gap-2 px-4 py-2.5 bg-claude-text text-white rounded-xl text-sm font-sans font-medium hover:bg-black transition-colors"
-                    >
-                      <Twitter size={16} />
-                      X
-                    </button>
-                    <button
-                      onClick={() => shareOnFacebook(selectedArticle)}
-                      className="flex items-center gap-2 px-4 py-2.5 bg-[#1877F2] text-white rounded-xl text-sm font-sans font-medium hover:bg-[#0D65D9] transition-colors"
-                    >
-                      <Facebook size={16} />
-                      Facebook
-                    </button>
-                    <button
-                      onClick={() => copyLink(selectedArticle)}
-                      className="flex items-center gap-2 px-4 py-2.5 bg-white border border-claude-border text-claude-text rounded-xl text-sm font-sans font-medium hover:bg-claude-bg transition-colors"
-                    >
-                      {copied ? <Check size={16} className="text-green-600" /> : <Link2 size={16} />}
-                      {copied ? 'Скопійовано!' : 'Копіювати посилання'}
-                    </button>
-                  </div>
-                </div>
-
-                {/* ── Comments Section ── */}
-                <div className="mt-8 pt-6 border-t border-claude-border">
-                  <div className="flex items-center gap-2 mb-6">
-                    <MessageSquare size={18} className="text-claude-accent" />
-                    <h3 className="text-base font-sans font-medium text-claude-text">
-                      Коментарі {comments.length > 0 && <span className="text-claude-subtext font-normal">({comments.length})</span>}
-                    </h3>
-                  </div>
-
-                  {/* Comment input */}
-                  {isAuthenticated ? (
-                    <div className="flex gap-3 mb-6">
-                      <div className="flex-shrink-0 w-9 h-9 rounded-full bg-claude-accent/10 flex items-center justify-center overflow-hidden">
-                        {user?.picture ? (
-                          <img src={user.picture} alt="" className="w-full h-full object-cover" />
-                        ) : (
-                          <span className="text-sm font-medium text-claude-accent">{(user?.name || user?.email || '?')[0].toUpperCase()}</span>
-                        )}
-                      </div>
-                      <div className="flex-1 flex gap-2">
-                        <input
-                          type="text"
-                          value={commentText}
-                          onChange={(e) => setCommentText(e.target.value)}
-                          onKeyDown={(e) => e.key === 'Enter' && !e.shiftKey && handleSubmitComment()}
-                          placeholder="Написати коментар..."
-                          maxLength={2000}
-                          className="flex-1 px-4 py-2.5 bg-white border border-claude-border rounded-xl text-sm text-claude-text placeholder-claude-subtext/50 focus:outline-none focus:ring-2 focus:ring-claude-accent/20 focus:border-claude-accent transition-all font-sans"
-                        />
-                        <button
-                          onClick={handleSubmitComment}
-                          disabled={!commentText.trim() || commentSubmitting}
-                          className="px-3 py-2.5 bg-claude-accent text-white rounded-xl hover:bg-[#C66345] transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                        >
-                          <Send size={16} />
-                        </button>
-                      </div>
-                    </div>
-                  ) : (
-                    <a
-                      href={`/login?returnUrl=${encodeURIComponent('/blog?article=' + selectedArticle.id)}`}
-                      className="flex items-center justify-center gap-2 mb-6 px-4 py-3 bg-claude-bg border border-claude-border rounded-xl text-sm text-claude-subtext hover:text-claude-text hover:border-claude-accent/30 transition-all font-sans"
-                    >
-                      <LogIn size={16} />
-                      Увійдіть, щоб залишити коментар
-                    </a>
-                  )}
-
-                  {/* Comments list */}
-                  {commentsLoading ? (
-                    <div className="text-center py-6 text-sm text-claude-subtext font-sans">Завантаження...</div>
-                  ) : comments.length === 0 ? (
-                    <div className="text-center py-6 text-sm text-claude-subtext font-sans">
-                      Поки немає коментарів. Будьте першим!
-                    </div>
-                  ) : (
-                    <div className="space-y-4">
-                      {comments.map((comment) => (
-                        <div key={comment.id} className="flex gap-3 group">
-                          <div className="flex-shrink-0 w-8 h-8 rounded-full bg-claude-accent/10 flex items-center justify-center overflow-hidden">
-                            {comment.user_picture ? (
-                              <img src={comment.user_picture} alt="" className="w-full h-full object-cover" />
-                            ) : (
-                              <span className="text-xs font-medium text-claude-accent">
-                                {(comment.user_name || '?')[0].toUpperCase()}
-                              </span>
-                            )}
-                          </div>
-                          <div className="flex-1 min-w-0">
-                            <div className="flex items-center gap-2">
-                              <span className="text-sm font-medium text-claude-text font-sans">
-                                {comment.user_name || 'Анонім'}
-                              </span>
-                              <span className="text-xs text-claude-subtext/60 font-sans">
-                                {new Date(comment.created_at).toLocaleDateString('uk-UA', {
-                                  day: 'numeric', month: 'short', year: 'numeric', hour: '2-digit', minute: '2-digit',
-                                })}
-                              </span>
-                              {isAuthenticated && user?.id === comment.user_id && (
-                                <button
-                                  onClick={() => handleDeleteComment(comment.id)}
-                                  className="opacity-0 group-hover:opacity-100 text-claude-subtext/40 hover:text-red-500 transition-all ml-auto"
-                                  title="Видалити"
-                                >
-                                  <Trash2 size={14} />
-                                </button>
-                              )}
-                            </div>
-                            <p className="text-sm text-claude-text font-sans mt-1 leading-relaxed">{comment.content}</p>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              </div>
-            </motion.div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </div>
-  );
-}
