+++ lexwebapp/src/pages/LoginPage/ForgotPasswordModal.tsx
@@ -1,0 +1,110 @@
+import { useState } from 'react';
+import { motion } from 'framer-motion';
+import { Mail, Loader2 } from 'lucide-react';
+import showToast from '../../utils/toast';
+
+const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:3000';
+const BASE_URL = API_URL.replace(/\/api$/, '');
+
+interface ForgotPasswordModalProps {
+  initialEmail: string;
+  onClose: () => void;
+}
+
+export function ForgotPasswordModal({ initialEmail, onClose }: ForgotPasswordModalProps) {
+  const [email, setEmail] = useState(initialEmail);
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const handleSubmit = async () => {
+    if (!email) {
+      setError('Please enter your email');
+      return;
+    }
+
+    setIsLoading(true);
+    setError(null);
+
+    try {
+      const response = await fetch(`${BASE_URL}/auth/forgot-password`, {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ email }),
+      });
+
+      const data = await response.json();
+
+      if (!response.ok) {
+        throw new Error(data.message || 'Failed to send reset email');
+      }
+
+      showToast.success('Password reset link sent to your email');
+      onClose();
+    } catch (err: unknown) {
+      const msg = err instanceof Error ? err.message : 'Failed to send reset email';
+      setError(msg);
+      showToast.error('Failed to send reset email');
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <motion.div
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      exit={{ opacity: 0 }}
+      className="fixed inset-0 bg-black/50 flex items-center justify-center z-50 p-4"
+      onClick={(e) => {
+        if (e.target === e.currentTarget) onClose();
+      }}
+    >
+      <motion.div
+        initial={{ opacity: 0, scale: 0.95, y: 20 }}
+        animate={{ opacity: 1, scale: 1, y: 0 }}
+        exit={{ opacity: 0, scale: 0.95, y: 20 }}
+        className="bg-white rounded-2xl p-8 max-w-md w-full shadow-2xl"
+      >
+        <h2 className="text-xl font-sans text-claude-text mb-2">Скидання паролю</h2>
+        <p className="text-sm text-claude-subtext mb-6">
+          Введіть email вашого акаунту, і ми відправимо вам посилання для скидання паролю.
+        </p>
+
+        {error && (
+          <p className="text-sm text-red-600 mb-4">{error}</p>
+        )}
+
+        <div className="relative mb-4">
+          <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+            <Mail size={18} className="text-claude-subtext" />
+          </div>
+          <input
+            type="email"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            onKeyDown={(e) => e.key === 'Enter' && handleSubmit()}
+            placeholder="your@email.com"
+            autoFocus
+            className="block w-full pl-10 pr-4 py-3 bg-white border border-claude-border rounded-xl text-claude-text placeholder-claude-subtext/50 focus:outline-none focus:ring-2 focus:ring-claude-accent/20 focus:border-claude-accent transition-all font-sans"
+          />
+        </div>
+
+        <div className="flex gap-3">
+          <button
+            onClick={onClose}
+            className="flex-1 px-4 py-3 border border-claude-border rounded-xl hover:bg-claude-bg transition-colors font-sans"
+          >
+            Скасувати
+          </button>
+          <button
+            onClick={handleSubmit}
+            disabled={isLoading}
+            className="flex-1 px-4 py-3 bg-black text-white rounded-xl hover:bg-gray-800 transition-colors font-sans disabled:opacity-50 flex items-center justify-center"
+          >
+            {isLoading ? <Loader2 size={18} className="animate-spin" /> : 'Відправити'}
+          </button>
+        </div>
+      </motion.div>
+    </motion.div>
+  );
+}
